# docs/swagger.yaml
basePath: /
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: エラーが発生しました
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Trihackathon API Server
  title: Trihackathon API
  version: "1.0"
paths:
...
  /cron/weekly-evaluation:
    post:
      description: 全activeチームの週次評価を実行し、HP更新・disbanded処理を行う
      parameters:
      - description: Cronシークレットキー
        in: header
        name: X-Cron-Secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 週次評価実行
      tags:
      - cron
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"